+++ src/openassetio-python/cmodule/src/managerApi/ManagerInterfaceBinding.cpp
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-// Copyright 2013-2022 The Foundry Visionmongers Ltd
+// Copyright 2013-2023 The Foundry Visionmongers Ltd
 #include <pybind11/functional.h>
 #include <pybind11/stl.h>
 
@@ -86,6 +86,11 @@ struct PyManagerInterface : ManagerInterface {
                                              const HostSessionPtr& hostSession) const override {
     PYBIND11_OVERRIDE_PURE(bool, ManagerInterface, isEntityReferenceString, someString,
                            hostSession);
+  }
+
+  [[nodiscard]] StrMap updateTerminology(StrMap terms,
+                                         const HostSessionPtr& hostSession) const override {
+    PYBIND11_OVERRIDE(StrMap, ManagerInterface, updateTerminology, std::move(terms), hostSession);
   }
 
   void resolve(const EntityReferences& entityReferences, const trait::TraitSet& traitSet,
@@ -202,6 +207,8 @@ void registerManagerInterface(const py::module& mod) {
            py::arg("token"), py::arg("hostSession").none(false))
       .def("isEntityReferenceString", &ManagerInterface::isEntityReferenceString,
            py::arg("someString"), py::arg("hostSession").none(false))
+      .def("updateTerminology", &ManagerInterface::updateTerminology, py::arg("terms"),
+           py::arg("hostSession").none(false))
       .def("resolve", &ManagerInterface::resolve, py::arg("entityReferences"), py::arg("traitSet"),
            py::arg("context").none(false), py::arg("hostSession").none(false),
            py::arg("successCallback"), py::arg("errorCallback"))